requirements.txt
aiocontextvars==0.2.2
astroid==2.3.3
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
contextvars==2.4
google-api-core==1.14.3
google-auth==1.8.1
google-cloud-bigquery==1.22.0
google-cloud-core==1.1.0
google-resumable-media==0.5.0
googleapis-common-protos==1.6.0
idna==2.8
immutables==0.11
isort==4.3.21
lazy-object-proxy==1.4.3
loguru==0.4.0
mccabe==0.6.1
numpy==1.17.4
pandas==0.25.3
pkg-resources==0.0.0
protobuf==3.11.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pylint==2.4.4
python-dateutil==2.8.1
python-dotenv==0.10.3
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.13.0
sodapy==2.0.0
typed-ast==1.4.0
urllib3==1.25.7
wrapt==1.11.2
